좌석 선택 및 결제 창 생성

+++ src/main/java/movie_ticket/FilmGo/controller/booking/dto/BookingRequestForm.java
@@ -16,5 +16,5 @@ public class BookingRequestForm {
     private Long movieId;
 
     @NotNull
-    private List<Integer> seatIds;
+    private List<Long> seatIds;
 }
